Refactors form validation logic
Moves validation logic into reusable utility functions and separate files for better organization and maintainability. This change improves code readability and testability.
The step validations are stored in stepValidations.ts file.
The types regarding FieldType, ValidationResult, FieldValidation and StepValidation are stored in types.ts file.
The validation utils such as getNestedValue, validateField and validateStep are stored in validation.ts file.
